+ packs util

+++ Nut/src/main/java/goals/packs/War.java
@@ -2,8 +2,11 @@ package nut.goals.packs;
 
 import nut.goals.GoalException;
 import nut.goals.packs.Zip;
+import nut.goals.packs.util.CopyFiles;
+import nut.goals.packs.util.ZipFiles;
 import nut.logging.Log;
 import nut.model.Dependency;
+import nut.model.Project;
 
 import java.io.File;
 import java.io.IOException;
@@ -22,45 +25,75 @@ public class War
 {
     /** Instance logger */
     private Log log;
-
     private String name;
-    private String repository;
-    private List<Dependency> dependencies;
 
     // ==========================================================================
-    public War( String name, String repository, List<Dependency> dependencies )
+    public War( String name )
     {
         this.name = name;
-        this.repository = repository;
-        this.dependencies = dependencies;
         log = new Log();
     }
 
     // ==========================================================================
-    public void archive( String directory ) throws GoalException
+    public void archive(Project project) throws GoalException
     {
+        String basedir              = project.getBaseDirectory();
+        String resourceDirectory    = basedir + File.separator + project.getResourceDirectory();
+        String outputDirectory      = basedir + File.separator + project.getOutputDirectory();
+        String webappDirectory      = basedir + File.separator + project.getWebappDirectory();
+
+        log.debug( "resourceDirectory   = " + resourceDirectory );
+        log.debug( "outputDirectory     = " + outputDirectory );
+        log.debug( "webappDirectory     = " + webappDirectory );
+
         // ----------------------------------------------------------------------
-        String libDirectory = directory + "/WEB-INF/lib";
+        // copy resource files
+        try {
+            CopyFiles resource = new CopyFiles( resourceDirectory, webappDirectory );
+            resource.process();
+        } catch( Exception e ) {
+            log.error( "Failed to copy resource: " + e.getMessage(), e );
+            throw new GoalException(e.getMessage());
+        }
+        // copy compiled files
+        try {
+            CopyFiles compiled = new CopyFiles( outputDirectory,   webappDirectory + "/WEB-INF/classes" );
+            compiled.process();
+        } catch( Exception e ) {
+            log.error( "Failed to copy compiled source: " + e.getMessage(), e );
+            throw new GoalException(e.getMessage());
+        }
+        // ----------------------------------------------------------------------
+        // copy dependencies
+        String libDirectory = webappDirectory + "/WEB-INF/lib";
         File libDir = new File( libDirectory );
         if ( !libDir.exists() ) {
             libDir.mkdirs();
         }
-
+        String repository = project.getRepository();
         try {
             log.debug("* copy dependencies to " + libDirectory);
-            for ( Iterator it = dependencies.iterator(); it.hasNext(); ) {
+            for ( Iterator it = project.getDependencies().iterator(); it.hasNext(); ) {
               Dependency dep = (Dependency) it.next();
               String depPath = repository + dep.getPath();
               log.debug("** " + depPath);
               Path dest = Paths.get(libDirectory + "/" + dep.getName());
-              Files.copy(Paths.get(depPath), dest);
+              if ( !dest.toFile().exists() )
+                  Files.copy(Paths.get(depPath), dest);
             }
         } catch( Exception e ) {
             log.error( "Failed to copy dependency: " + e.getMessage(), e );
             throw new GoalException(e.getMessage());
         }
-        Zip zip = new Zip(this.name);
-        zip.archive(directory);
+        // ----------------------------------------------------------------------
+        // Create war file
+        try {
+            ZipFiles zip = new ZipFiles( webappDirectory, name );
+            zip.process();
+        } catch( Exception e ) {
+            log.error( "Failed to zip: " + e.getMessage(), e );
+            throw new GoalException(e.getMessage());
+        }
     }
 
     // ==========================================================================
